move unit tests out of the unit folder so that i can run coverage just against it
