+++ java/src/test/java/com/saucelabs/simplesauce/EdgeTests.java
@@ -1,4 +1,4 @@
-package com.saucelabs.simplesauce.unit;
+package com.saucelabs.simplesauce;
 
 import org.junit.Test;
