+++ java/src/test/java/com/saucelabs/simplesauce/SauceSessionTest.java
@@ -1,4 +1,4 @@
-package com.saucelabs.simplesauce.unit;
+package com.saucelabs.simplesauce;
 
 import com.saucelabs.simplesauce.ConcreteRemoteDriver;
 import com.saucelabs.simplesauce.DataCenter;
