+++ java/src/test/java/com/saucelabs/simplesauce/WindowsTests.java
@@ -1,4 +1,4 @@
-package com.saucelabs.simplesauce.unit;
+package com.saucelabs.simplesauce;
 
 import org.junit.Test;
